# ClientApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Service_API",
    "version": "v1"
  },
  "paths": {
    "/getall": {
      "get": {
        "tags": [
          "ProductsInfoe"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductsInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/getone/{id}": {
      "get": {
        "tags": [
          "ProductsInfoe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsInfo"
                }
              }
            }
          }
        }
      }
    },
    "/update/{id}": {
      "put": {
        "tags": [
          "ProductsInfoe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/create": {
      "post": {
        "tags": [
          "ProductsInfoe"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsInfo"
                }
              }
            }
          }
        }
      }
    },
    "/delete/{id}": {
      "delete": {
        "tags": [
          "ProductsInfoe"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductsInfo": {
        "type": "object",
        "properties": {
          "ProductRowId": {
            "type": "integer",
            "format": "int32"
          },
          "ProductId": {
            "type": "string",
            "nullable": true
          },
          "ProductName": {
            "type": "string",
            "nullable": true
          },
          "CategoryName": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "Manufacturer": {
            "type": "string",
            "nullable": true
          },
          "Price": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}